test: extract reified namespace suites
